stacking

+++ getline.c
@@ -32,7 +32,7 @@ ssize_t getline(char **lineptr, size_t *n, FILE *stream)
 	p = bufptr;
 	while (c != EOF)
 	{
-		if ((p - bufptr) > (size - 1))
+		if ((p - bufptr) > (long)(size - 1))
 		{
 			size = size + 128;
 			bufptr = realloc(bufptr, size);
